집계 수정

+++ src/main/java/com/example/wonjang/controller/AdminController.java
@@ -51,7 +51,7 @@ public class AdminController {
             Model model
             , @CurrentUser Member member
     ){
-        Map<DayOfWeek, Long> thisWeekVisits = userVisitService.getThisWeekVisits();
+        Map<DayOfWeek, Integer> thisWeekVisits = userVisitService.getThisWeekVisits();
         model.addAttribute("thisWeekVisits",thisWeekVisits);
         return "admin/index";
     }
